# lab2/payment-auth-system/docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.AuthorizeTransactionRequest:
    properties:
      amount:
        type: integer
      card_number:
        type: string
      terminal_serial_number:
        type: string
    type: object
  handlers.AuthorizeTransactionResponse:
    properties:
      balance_left:
        type: integer
      card_id:
        type: integer
      message:
        type: string
      status:
        type: string
      terminal_id:
        type: integer
    type: object
...
  handlers.CreateTransactionRequest:
    properties:
      amount:
        type: integer
      card_id:
        type: integer
      terminal_id:
        type: integer
    type: object
...
  handlers.UpdateTransactionRequest:
    properties:
      amount:
        type: integer
      card_id:
        type: integer
      terminal_id:
        type: integer
    type: object
...
  models.Transaction:
    properties:
      amount:
        format: int64
        type: integer
      cardID:
        format: int64
        type: integer
      createdAt:
        type: string
      id:
        format: int64
        type: integer
      terminalID:
        format: int64
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API сервера авторизации платежей транспортными картами
  title: Payment Auth System API
  version: "1.0"
paths:
...
  /terminal/authorize:
    post:
      consumes:
      - application/json
      description: Проверяет карту, терминал и баланс, затем создаёт транзакцию
      parameters:
      - description: Authorize transaction request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AuthorizeTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AuthorizeTransactionResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.AuthorizeTransactionResponse'
      summary: Авторизация платежа терминалом
      tags:
      - terminal
...
  /transactions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Transaction'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получить список транзакций
      tags:
      - transactions
    post:
      consumes:
      - application/json
      parameters:
      - description: Create transaction request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Создать транзакцию
      tags:
      - transactions
  /transactions/{id}:
    delete:
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Удалить транзакцию
      tags:
      - transactions
    get:
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Transaction'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получить транзакцию по ID
      tags:
      - transactions
    put:
      consumes:
      - application/json
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update transaction request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Обновить транзакцию
      tags:
      - transactions
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"